fix(invitations): retain names for newly invited accounts
