+++ docs/openapi/openapi.yaml
@@ -428,6 +428,28 @@ definitions:
         additionalProperties: {}
         type: object
     type: object
+  dev.createInvitationRequest:
+    properties:
+      email:
+        type: string
+      name:
+        type: string
+      role:
+        type: string
+    type: object
+  dev.createInvitationResponse:
+    properties:
+      email:
+        type: string
+      expires_at:
+        type: string
+      invitation_id:
+        type: string
+      invite_link:
+        type: string
+      role:
+        type: string
+    type: object
   dev.createModelBody:
     properties:
       adapter:
@@ -3626,8 +3648,9 @@ paths:
     post:
       consumes:
       - application/json
-      description: New accounts supply a password. Existing accounts must be signed
-        in as the invited user; their password is never changed.
+      description: New accounts supply a password and use the saved invitation name
+        when provided. Existing accounts must be signed in as the invited user; their
+        name and password are never changed.
       parameters:
       - description: Invitation and optional new-account password
         in: body
@@ -7211,6 +7234,64 @@ paths:
       summary: Resolve a pending approval or user question
       tags:
       - interactions
+  /api/v1/workspaces/{workspaceID}/invitations:
+    post:
+      consumes:
+      - application/json
+      description: Stores the optional name for new-account creation. Existing account
+        names are preserved.
+      parameters:
+      - description: Workspace UUID
+        in: path
+        name: workspaceID
+        required: true
+        type: string
+      - description: Invitation details
+        in: body
+        name: body
+        required: true
+        schema:
+          $ref: '#/definitions/dev.createInvitationRequest'
+      produces:
+      - application/json
+      responses:
+        "201":
+          description: Created
+          schema:
+            $ref: '#/definitions/dev.createInvitationResponse'
+        "400":
+          description: Invalid invitation details
+          schema:
+            additionalProperties:
+              type: string
+            type: object
+        "401":
+          description: Authentication required
+          schema:
+            additionalProperties:
+              type: string
+            type: object
+        "403":
+          description: Invitation not allowed
+          schema:
+            additionalProperties:
+              type: string
+            type: object
+        "409":
+          description: Invitation already pending
+          schema:
+            additionalProperties:
+              type: string
+            type: object
+        "500":
+          description: Invitation creation failed
+          schema:
+            additionalProperties:
+              type: string
+            type: object
+      summary: Create a workspace invitation
+      tags:
+      - invitations
   /api/v1/workspaces/{workspaceID}/mcp-directory:
     get:
       parameters:
