+++ server/internal/dev/routes_workspace_members.go
@@ -14,7 +14,6 @@ import (
 	authinvite "github.com/MiniMax-AI-Dev/parsar/server/internal/auth/invite"
 	"github.com/MiniMax-AI-Dev/parsar/server/internal/store"
 	"github.com/go-chi/chi/v5"
-	"github.com/google/uuid"
 )
 
 // listWorkspaceMembers lists members of a workspace.
@@ -130,20 +129,6 @@ func addWorkspaceMember(runtimeStore RuntimeStore) http.HandlerFunc {
 
 // ── Invitation handlers ─────────────────────────────────────────
 
-type createInvitationRequest struct {
-	Email string `json:"email"`
-	Name  string `json:"name,omitempty"`
-	Role  string `json:"role"`
-}
-
-type createInvitationResponse struct {
-	InvitationID string `json:"invitation_id"`
-	InviteLink   string `json:"invite_link"`
-	Email        string `json:"email"`
-	Role         string `json:"role"`
-	ExpiresAt    string `json:"expires_at"`
-}
-
 func invitationCallerRole(r *http.Request, runtimeStore RuntimeStore, workspaceID string) (string, string, error) {
 	ctx := requestContextForRBAC(r)
 	callerID := auth.UserIDFromContext(ctx)
@@ -155,77 +140,6 @@ func invitationCallerRole(r *http.Request, runtimeStore RuntimeStore, workspaceI
 		return "", "", err
 	}
 	return callerID, role, nil
-}
-
-func createInvitation(runtimeStore RuntimeStore, cfg *routerConfig) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		workspaceID := strings.TrimSpace(chi.URLParam(r, "workspaceID"))
-		if !isUUID(workspaceID) {
-			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "workspace_id must be a valid uuid"})
-			return
-		}
-		if err := requireWorkspaceMemberNotViewer(r, runtimeStore, workspaceID); err != nil {
-			writeRBACError(w, err)
-			return
-		}
-		var req createInvitationRequest
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid json body"})
-			return
-		}
-		req.Email = strings.TrimSpace(req.Email)
-		req.Role = strings.TrimSpace(req.Role)
-		if req.Email == "" {
-			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "email is required"})
-			return
-		}
-		if !store.IsValidMemberRole(req.Role) {
-			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "role must be one of owner|admin|member|viewer"})
-			return
-		}
-
-		now := time.Now().UTC()
-		expiresAt := now.Add(authinvite.MaxLifetime)
-		invID := uuid.New().String()
-		token := invID
-
-		callerID, callerRole, err := invitationCallerRole(r, runtimeStore, workspaceID)
-		if err != nil {
-			writeRBACError(w, err)
-			return
-		}
-		if callerRole == "member" && req.Role != "member" {
-			writeJSON(w, http.StatusForbidden, map[string]string{"error": "members can only invite new members with the member role"})
-			return
-		}
-
-		if err := runtimeStore.CreateInvitation(r.Context(), store.CreateInvitationInput{
-			ID:          invID,
-			TokenHash:   authinvite.TokenHash(token),
-			WorkspaceID: workspaceID,
-			Email:       req.Email,
-			Role:        req.Role,
-			InvitedBy:   callerID,
-			ExpiresAt:   expiresAt,
-			Now:         now,
-		}); err != nil {
-			if strings.Contains(err.Error(), "uk_workspace_invitations_pending_email") {
-				writeJSON(w, http.StatusConflict, map[string]string{"error": "an invitation is already pending for this email"})
-				return
-			}
-			writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to create invitation"})
-			return
-		}
-
-		link := cfg.publicURL + "/invite/" + token
-		writeJSON(w, http.StatusCreated, createInvitationResponse{
-			InvitationID: invID,
-			InviteLink:   link,
-			Email:        store.NormalizeEmail(req.Email),
-			Role:         req.Role,
-			ExpiresAt:    expiresAt.Format(time.RFC3339),
-		})
-	}
 }
 
 type pendingInvitationResponse struct {
