+++ server/internal/store/invitation_accept_test.go
@@ -19,26 +19,31 @@ func TestAcceptInvitationPreservesExistingAccount(t *testing.T) {
 		t.Fatal(err)
 	}
 	now := time.Now().UTC()
-	invite := func(workspace, email, hash string) AcceptInvitationInput {
+	invite := func(workspace, email, name, hash string) AcceptInvitationInput {
 		t.Helper()
 		token := sha256.Sum256([]byte(newID()))
 		in := AcceptInvitationInput{TokenHash: token[:], Email: email, Role: "member", WorkspaceID: workspace, PasswordHash: hash, Now: now}
-		err := st.CreateInvitation(ctx, CreateInvitationInput{ID: newID(), TokenHash: in.TokenHash, WorkspaceID: workspace, Email: email, Role: in.Role, InvitedBy: ids.UserID, ExpiresAt: now.Add(time.Hour), Now: now})
+		err := st.CreateInvitation(ctx, CreateInvitationInput{ID: newID(), TokenHash: in.TokenHash, WorkspaceID: workspace, Email: email, Name: name, Role: in.Role, InvitedBy: ids.UserID, ExpiresAt: now.Add(time.Hour), Now: now})
 		if err != nil {
 			t.Fatal(err)
 		}
+		saved, err := st.GetInvitationByTokenHash(ctx, in.TokenHash)
+		if err != nil {
+			t.Fatal(err)
+		}
+		in.Name = saved.Name
 		return in
 	}
-	first := invite(ids.WorkspaceID, "invitee@example.com", "original-hash")
+	first := invite(ids.WorkspaceID, "invitee@example.com", " QA 员工 ", "original-hash")
 	created, err := st.AcceptInvitation(ctx, first)
-	if err != nil || !created.UserCreated {
+	if err != nil || !created.UserCreated || created.Member.UserName != "QA 员工" {
 		t.Fatalf("new account: %+v, %v", created, err)
 	}
 	target, err := st.CreateWorkspace(ctx, CreateWorkspaceInput{Name: "Invitation target", CreatedBy: ids.UserID, Now: now})
 	if err != nil {
 		t.Fatal(err)
 	}
-	second := invite(target.Workspace.ID, first.Email, "replacement-hash")
+	second := invite(target.Workspace.ID, first.Email, "Replacement name", "replacement-hash")
 	q := sqlc.New(db)
 	assertUnchanged := func() {
 		t.Helper()
@@ -64,7 +69,7 @@ func TestAcceptInvitationPreservesExistingAccount(t *testing.T) {
 	}
 	second.ActorUserID = created.Member.UserID
 	accepted, err := st.AcceptInvitation(ctx, second)
-	if err != nil || accepted.UserCreated || accepted.Member.UserID != created.Member.UserID {
+	if err != nil || accepted.UserCreated || accepted.Member.UserID != created.Member.UserID || accepted.Member.UserName != created.Member.UserName {
 		t.Fatalf("authenticated acceptance: %+v, %v", accepted, err)
 	}
 	assertUnchanged()
@@ -72,13 +77,13 @@ func TestAcceptInvitationPreservesExistingAccount(t *testing.T) {
 		t.Fatalf("replay: %v", err)
 	}
 
-	third := invite(target.Workspace.ID, "missing-password@example.com", "")
+	third := invite(target.Workspace.ID, "missing-password@example.com", "", "")
 	third.ActorUserID = ids.UserID
 	if _, err := st.AcceptInvitation(ctx, third); !errors.Is(err, ErrInvalidInput) {
 		t.Fatalf("missing new-account password: %v", err)
 	}
 	third.PasswordHash = "new-account-hash"
-	if result, err := st.AcceptInvitation(ctx, third); err != nil || !result.UserCreated {
+	if result, err := st.AcceptInvitation(ctx, third); err != nil || !result.UserCreated || result.Member.UserName != "missing-password" {
 		t.Fatalf("new account retry: %+v, %v", result, err)
 	}
 }
