+++ server/internal/store/store.go
@@ -5147,17 +5147,6 @@ func (s *Store) AddWorkspaceMember(ctx context.Context, input AddWorkspaceMember
 
 // ── Invitation CRUD ─────────────────────────────────────────────
 
-type CreateInvitationInput struct {
-	ID          string
-	TokenHash   []byte
-	WorkspaceID string
-	Email       string
-	Role        string
-	InvitedBy   string
-	ExpiresAt   time.Time
-	Now         time.Time
-}
-
 type PendingInvitationRead struct {
 	ID            string    `json:"id"`
 	TokenHash     []byte    `json:"-"`
@@ -5167,33 +5156,6 @@ type PendingInvitationRead struct {
 	InvitedByName string    `json:"invited_by_name"`
 	ExpiresAt     time.Time `json:"expires_at"`
 	CreatedAt     time.Time `json:"created_at"`
-}
-
-type InvitationRead struct {
-	ID            string     `json:"id"`
-	WorkspaceID   string     `json:"workspace_id"`
-	Email         string     `json:"email"`
-	Role          string     `json:"role"`
-	InvitedBy     string     `json:"invited_by"`
-	ExpiresAt     time.Time  `json:"expires_at"`
-	AcceptedAt    *time.Time `json:"accepted_at"`
-	RevokedAt     *time.Time `json:"revoked_at"`
-	CreatedAt     time.Time  `json:"created_at"`
-	WorkspaceName string     `json:"workspace_name"`
-}
-
-func (s *Store) CreateInvitation(ctx context.Context, input CreateInvitationInput) error {
-	q := sqlc.New(s.db)
-	return q.CreateWorkspaceInvitation(ctx, sqlc.CreateWorkspaceInvitationParams{
-		ID:          mustUUID(input.ID),
-		TokenHash:   input.TokenHash,
-		WorkspaceID: mustUUID(input.WorkspaceID),
-		Email:       normalizeEmail(input.Email),
-		Role:        input.Role,
-		InvitedBy:   mustUUID(input.InvitedBy),
-		ExpiresAt:   timestamptz(input.ExpiresAt),
-		CreatedAt:   timestamptz(input.Now),
-	})
 }
 
 func (s *Store) ListPendingInvitations(ctx context.Context, workspaceID string) ([]PendingInvitationRead, error) {
@@ -5276,31 +5238,6 @@ func (s *Store) RevokeOwnInvitation(ctx context.Context, workspaceID, invitation
 		InvitedBy:   mustUUID(invitedBy),
 		Now:         timestamptz(time.Now().UTC()),
 	})
-}
-
-func (s *Store) GetInvitationByTokenHash(ctx context.Context, tokenHash []byte) (InvitationRead, error) {
-	q := sqlc.New(s.db)
-	row, err := q.GetWorkspaceInvitationByTokenHash(ctx, tokenHash)
-	if err != nil {
-		return InvitationRead{}, err
-	}
-	inv := InvitationRead{
-		ID:            row.ID,
-		WorkspaceID:   row.WorkspaceID,
-		Email:         row.Email,
-		Role:          row.Role,
-		InvitedBy:     row.InvitedBy,
-		ExpiresAt:     row.ExpiresAt.Time,
-		CreatedAt:     row.CreatedAt.Time,
-		WorkspaceName: row.WorkspaceName,
-	}
-	if row.AcceptedAt.Valid {
-		inv.AcceptedAt = &row.AcceptedAt.Time
-	}
-	if row.RevokedAt.Valid {
-		inv.RevokedAt = &row.RevokedAt.Time
-	}
-	return inv, nil
 }
 
 // UpdateWorkspaceMemberRole flips an existing member's role. Returns
